아키텍처 패턴 나누기
